osdu-app-with-connector/requirements.txt
httpx>=0.27
azure-identity>=1.15
pydantic>=2.5
PyYAML>=6
tenacity>=8
